misc/HtmlTestSuite/results/test-0333.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Tests that overflowing the inset-modified containing block triggers position fallback</title>
<link href="https://drafts.csswg.org/css-anchor-1/#fallback" rel="help"/>
<link href="mailto:xiaochengh@chromium.org" rel="author"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<style>
.cb {
  width: 400px;
  height: 400px;
  transform: scale(1);
  background: yellow;
}
.anchor1 {
  anchor-name: --a;
  margin-left: 100px;
  width: 100px;
  height: 100px;
  background: blue;
}
.target {
  position: absolute;
  position-fallback: --fallback;
  width: min-content;
  height: 100px;
  background: orange;
}
.inline-spacer {
  display: inline-block;
  width: 200px;
  height: 100px;
}
@position-fallback --fallback {
  @try { /* 1: Position to the left of the anchor. */
    left: 0;
    right: anchor(--a left);
    top: anchor(--a top);
  }
  @try { /* 2: Position to the right of the anchor. */
    left: anchor(--a right);
    right: 0;
    top: anchor(--a top);
  }
  @try { /* 3: Placeholder fallback that shouldn't be selected when the previous
               ones do not overflow the available space. */
    inset: 0;
  }
}
</style>
</head><body onload="checkLayout('.target')">
  <!-- The first `@try` overflows the inset-modifed containing block -->
  <div class="cb">
    <div class="anchor1"/>
    <div class="target" data-expected-height="100" data-expected-width="200" data-offset-x="200" data-offset-y="0">
      <span class="inline-spacer"/>
    </div>
  </div>
</body>
</html>
